# Removed auto-decoding from std.conv.parse (part 2)