Benchmark code here: https://gist.github.com/JackStouffer/af53d497532a864f56a5b53cb35cd438

Gives about a 40% speed increase:

```
# old
$ ldc2 -O5 -release -boundscheck=off test.d && ./test
Total time: 5 secs, 266 ms, 64 μs, and 3 hnsecs

# new
$ ldc2 -O5 -release -boundscheck=off test.d && ./test
Total time: 3 secs, 295 ms, 663 μs, and 7 hnsecs
```