@9il 2016-07-28
@JackStouffer  please add test with global `__gshared char[]` integer number and `pragma(inline, false)` for parse functions. 4x-5x is large number, so,  we need to be sure that LDC optimisation does not brake the test.