@wilzbach 2016-07-28
> @JackStouffer please add test with global __gshared char[] integer number and pragma(inline, false) for parse functions. 4x-5x is large number, so, we need to be sure that LDC optimisation does not brake the test.

IIRC @klickverbot pointed out that a good way to prevent LDC optimizations is to use separate files & annotate the tested methods with `extern`.

I have a [small repo](https://github.com/wilzbach/d-benchmarks) for benchmark scripts which are separately compiled and then linked, maybe that helps you?