@9il 2016-07-28
btw, `pragma(inline, false)` works correctly only in the recent LDC alpha, or master