@JackStouffer 2016-07-31
Rebased.

Will add extra test when I get time.