@JackStouffer 2016-08-08
@9il I updated the benchmark code here to your requests: https://github.com/dlang/phobos/pull/4597