@andralex 2016-08-30
This doesn't look right. Most code should stay the same. The only change is at the top level, where decodable strings should be replaced with `byChar` or `.representation`.