@andralex 2016-08-30
@JackStouffer then just work on a temporary and in the end assign the reference from it. It's a classic.