@JackStouffer 2016-09-07
I don't understand. Operate on a temporary and reassign it how?

If I make a copy of `p` inside of `parse` and then use `byCodeUnit` on that, the copy of `p` is copied again by `byCodeUnit` and that is advanced and not the copy of `p`. So I would still have to manually `popFront` the copy in order to reassign it to `p`.