@9il 2016-10-06
@JackStouffer it is save to reassing ubyte array to string after intefer parsing. Please implement Andrei's suggestions