@JackStouffer 2016-10-12
I will get back to this eventually. I'm a bit swamped with school and work right now, so maybe around Thanksgiving.