@JackStouffer 2017-01-04
Closed in favor of https://github.com/dlang/phobos/pull/5015